completed/corrected french data parsing

+++ src/main/kotlin/org/soframel/health/covid/model/french/FrenchCovidDailyData.kt
@@ -10,32 +10,32 @@ class FrenchCovidDailyData constructor( ) {
     val nom: String=""
     val code: String=""
 
-    val casConfirmes: Int=0
-    val deces: Int=0
-    val decesEhpad: Int=0
-    val hospitalises: Int=0
-    val hospitalise: Int=0
-    val hospitalisation: Int=0
-    val reanimation: Int=0
-    val reanimations: Int=0
-    val gueris: Int=0
-    val casConfirmesEhpad: Int=0
-    val casEhpad:Int=0
-    val casPossiblesEhpad:Int=0
+    val casConfirmes: Long=0
+    val deces: Long=0
+    val decesEhpad: Long=0
+    val hospitalises: Long=0
+    val hospitalise: Long=0
+    val hospitalisation: Long=0
+    val reanimation: Long=0
+    val reanimations: Long=0
+    val gueris: Long=0
+    val casConfirmesEhpad: Long=0
+    val casEhpad:Long=0
+    val casPossiblesEhpad:Long=0
 
-    val nouvellesHospitalisations: Int=0
-    val nouvellesReanimations: Int=0
-    val hospitalisesReadaptation:Int=0
-    val hospitalisesAuxUrgences:Int=0
-    val hospitalisesConventionnelle:Int=0
-    val paysTouches: Int=0
+    val nouvellesHospitalisations: Long=0
+    val nouvellesReanimations: Long=0
+    val hospitalisesReadaptation:Long=0
+    val hospitalisesAuxUrgences:Long=0
+    val hospitalisesConventionnelle:Long=0
+    val paysTouches: Long=0
     val victimes= ArrayList<Victime>()
-    val capaciteLitsReanimation: Int=0
-    val capaciteTotaleLitsDisponibles:Int=0
-    val capaciteLitsSoinsContinus:Int=0
-    val capaciteLitsSoinsIntensifs:Int=0
-    val capaciteReanimation:Int=0
-    val depistes:Int=0
+    val capaciteLitsReanimation: Long=0
+    val capaciteTotaleLitsDisponibles:Long=0
+    val capaciteLitsSoinsContinus:Long=0
+    val capaciteLitsSoinsIntensifs:Long=0
+    val capaciteReanimation:Long=0
+    val depistes:Long=0
 
     override fun toString(): String {
         return "MinistereSanteData(date=$date, source=$source, sourceType=$sourceType, nom='$nom', code='$code', casConfirmes=$casConfirmes, deces=$deces, decesEhpad=$decesEhpad, hospitalises=$hospitalises, reanimation=$reanimation, gueris=$gueris, casConfirmesEhpad=$casConfirmesEhpad, nouvellesHospitalisations=$nouvellesHospitalisations, nouvellesReanimations=$nouvellesReanimations)"
